+++ src/ultra/printer.clj
@@ -3,5 +3,7 @@
             [whidbey.render :as render]))
 
 (defn cprint
+  "Puget's cprint, set to always use the Puget options."
+  {:added "0.1.3"}
   [x]
   (printer/cprint x render/puget-options))