+++ test/ultra/test/diff_test.clj
@@ -1,0 +1,19 @@
+(ns ultra.test.diff-test
+  (:require [clojure.test :refer :all]
+            [ultra.test.diff :as diff]))
+
+(deftest list-diff-works
+  (is (= (diff/list-diff [1 2] [2 1])
+         "lists seem to contain the same items, but with different ordering."))
+  (is (= (diff/list-diff [1 1] [1])
+         "some duplicate items in actual are not in expected."))
+  (is (= (diff/list-diff [1] [1 1])
+         "some duplicate items in expected are not in actual."))
+  (is (= (diff/list-diff [1] [1 2])
+         "actual is longer than expected."))
+  (is (= (diff/list-diff [1 2] [1])
+         "expected is longer than actual.")))
+
+(deftest list-diff-works-with-lists
+  (is (= (diff/list-diff (list 1 2) (list 2 1))
+         "lists seem to contain the same items, but with different ordering.")))